+++ modules/docker_artifacts_registry/main.tf
@@ -1,17 +1,21 @@
-resource "google_artifact_registry_repository" "my-repo" {
+locals {
+  delete_older_than_seconds = "${var.delete_older_than_days * 86400}s"
+}
+
+resource "google_artifact_registry_repository" "this" {
   location               = var.location
   project                = var.project_id
   repository_id          = var.artifacts_repository_id
   description            = var.description
   format                 = "DOCKER"
-  cleanup_policy_dry_run = false
+  cleanup_policy_dry_run = var.cleanup_policy_dry_run
   cleanup_policies {
-    id     = "delete-after-a-month"
+    id     = "delete-old-tagged"
     action = "DELETE"
     condition {
       tag_state    = "TAGGED"
       tag_prefixes = var.tags_to_delete_after_a_month
-      older_than   = "2592000s"
+      older_than   = local.delete_older_than_seconds
     }
   }
   cleanup_policies {