+++ modules/docker_artifacts_registry/outputs.tf
@@ -1,8 +1,9 @@
 output "repository_id" {
-  value       = google_artifact_registry_repository.my-repo.id
+  value       = google_artifact_registry_repository.this.id
   description = "Id of the repository"
 }
+
 output "repository_name" {
-  value       = google_artifact_registry_repository.my-repo.name
-  description = "name of the repository"
+  value       = google_artifact_registry_repository.this.name
+  description = "Name of the repository"
 }